requirements.txt
mss
Pillow
pyttsx3
ollama